Bug 973449 - [settings] refactor Do not track panel with AMD pattern
