+++ apps/settings/elements/do_not_track.html
@@ -50,6 +50,6 @@
       </ul>
     </div>
 
-    <script src="js/do_not_track.js"></script>
+    <panel data-path="panels/do_not_track/panel"></panel>
   </template>
 </element>
